Moves BpfAllocator from bpf_loader into program_runtime.
