+++ program-runtime/src/invoke_context.rs
@@ -109,24 +109,63 @@ impl<'a> ContextObject for InvokeContext<'a> {
     }
 }
 
-/// Based loosely on the unstable std::alloc::Alloc trait
-pub trait Alloc {
-    fn alloc(&mut self, layout: Layout) -> Result<u64, AllocErr>;
-    fn dealloc(&mut self, addr: u64, layout: Layout);
-}
-
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct AllocErr;
-
 impl fmt::Display for AllocErr {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.write_str("Error: Memory allocation failed")
     }
 }
 
+#[derive(Debug)]
+pub struct BpfAllocator {
+    #[allow(dead_code)]
+    heap: AlignedMemory<HOST_ALIGN>,
+    start: u64,
+    len: u64,
+    pos: u64,
+}
+
+impl BpfAllocator {
+    pub fn new(heap: AlignedMemory<HOST_ALIGN>, virtual_address: u64) -> Self {
+        let len = heap.len() as u64;
+        Self {
+            heap,
+            start: virtual_address,
+            len,
+            pos: 0,
+        }
+    }
+
+    pub fn heap_mut(&mut self) -> &mut AlignedMemory<HOST_ALIGN> {
+        &mut self.heap
+    }
+
+    pub fn alloc(&mut self, layout: Layout) -> Result<u64, AllocErr> {
+        let bytes_to_align = (self.pos as *const u8).align_offset(layout.align()) as u64;
+        if self
+            .pos
+            .saturating_add(layout.size() as u64)
+            .saturating_add(bytes_to_align)
+            <= self.len
+        {
+            self.pos += bytes_to_align;
+            let addr = self.start + self.pos;
+            self.pos += layout.size() as u64;
+            Ok(addr)
+        } else {
+            Err(AllocErr)
+        }
+    }
+
+    pub fn dealloc(&mut self, _addr: u64, _layout: Layout) {
+        // It's a bump allocator, free not supported
+    }
+}
+
 pub struct SyscallContext {
     pub stack: AlignedMemory<{ HOST_ALIGN }>,
-    pub allocator: Rc<RefCell<dyn Alloc>>,
+    pub allocator: Rc<RefCell<BpfAllocator>>,
     pub orig_account_lengths: Vec<usize>,
     pub trace_log: Vec<[u64; 12]>,
 }
